Убирает лишние доработки
